requirements.txt

python-telegram-bot>=20.0
flask>=3.0.0

gspread==6.1.4
oauth2client==4.1.3
telegram
google-api-python-client
google-auth

python-dotenv
psycopg2-binary
syncer
flask
google-api-python-client
google-auth
gspread
oauth2client
psycopg2-binary
python-dotenv
python-telegram-bot
